move constants and types to separate files
